requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
Flask-Login==0.6.2
PyJWT==2.8.0
passlib==1.7.4
python-dotenv==1.0.0